pyproject.toml
[tool.poetry]
name = "kg-poc"
version = "0.1.0"
description = ""
authors = ["aymanehachcham <aymanehachchaming@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
jupyter = "^1.0.0"
pandas = "^2.2.2"
chembl-downloader = "^0.4.5"
rdflib = "^7.0.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
